Created a 'microcore' of CAM, which captures the essence of metadata represented in tabular format.

+++ Source/CILAssemblyManipulator/CILAssemblyManipulator.Physical/IO/Defaults/Serialization.API.cs
@@ -684,7 +684,7 @@ public static partial class E_CILPhysical
       return infos.Select( info =>
       {
          MetaDataTable tbl;
-         return info != null && md.TryGetByTable( info.Table, out tbl ) ?
+         return info != null && md.TryGetByTable( (Int32) info.Table, out tbl ) ?
             tbl.RowCount :
             0;
       } ).ToArrayProxy().CQ;
